+++ resources/lang/de/reporting.php
@@ -6,5 +6,6 @@ return [
     'bank-deposits' => 'Bank: Einzahlungen',
     'kitchen' => 'Küche',
     'user-privileges' => 'Benutzerberechtigungen',
-    'role-privileges' => ' Rollenberechtigungen'
+    'role-privileges' => ' Rollenberechtigungen',
+    'user-access-sensitive-data' => 'Benutzer mit Zugang zu sensitiven Daten',
 ];