+++ resources/lang/en/reporting.php
@@ -6,5 +6,6 @@ return [
     'bank-deposits' => 'Bank: Deposits',
     'kitchen' => 'Kitchen',
     'user-privileges' => 'User privileges',
-    'role-privileges' => 'Role privileges'
+    'role-privileges' => 'Role privileges',
+    'user-access-sensitive-data' => 'Users with access to sensitive data',
 ];